Convert the component to a functional one
