+++ frontend/src/metabase-types/store/mocks/qb.ts
@@ -27,6 +27,7 @@ export const createMockQueryBuilderUIControlsState = (
   datasetEditorTab: "query",
   isShowingNotebookNativePreview: false,
   notebookNativePreviewSidebarWidth: null,
+  showSidebarTitle: false,
   ...opts,
 });
