+++ frontend/src/metabase/visualizations/components/ChartSettings/index.ts
@@ -1,0 +1,1 @@
+export * from "./ChartSettings";
